Medição de GPU arm

+++ app/src/main/java/com/example/newbenchmarking/interfaces/InferenceResult.kt
@@ -2,7 +2,7 @@ package com.example.newbenchmarking.interfaces
 
 data class InferenceResult(
     val cpuAverage: Float,
-    val gpuAverage: Float,
+    val gpuAverage: Int,
     val ramConsumedAverage: Float,
     val inferenceTimeAverage: Long,
     val loadTime: Long,
